refactor: clientで使用するコードのみexportするように変更
