+++ apps/server/src/index.ts
@@ -1,12 +1,3 @@
-import { type AppRouteType, app } from "@server/routes";
-import { type InferResponseType, hc } from "hono/client";
-
-export const createHonoClient = (
-  ...args: Parameters<typeof hc<AppRouteType>>
-): ReturnType<typeof hc<AppRouteType>> => {
-  return hc<AppRouteType>(...args);
-};
-
-export type InferHonoType<T> = InferResponseType<T>;
+import { app } from "@server/routes";
 
 export default app;
